+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.15.0] - 2026-05-13
+
+### Changed
+- `semantic`: restrict inequality operators `<, >, <=, >=` to numeric types (ints and floats)
+- `tests`: add IEEE-754 testing for native float comparisons testing NaN semantics
+
+---
+
 ## [1.14.0] - 2026-05-13
 
 ### Changed